+++ LivestockBazaar/PetFeatures.cs
@@ -1,0 +1,110 @@
+using StardewModdingAPI;
+using StardewValley;
+using StardewValley.GameData.Pets;
+using StardewValley.Internal;
+using StardewValley.Objects;
+
+namespace LivestockBazaar;
+
+internal static class PetFeatures
+{
+    internal const string ItemQuery_PET_ADOPTION = $"{ModEntry.ModId}_PET_ADOPTION";
+
+    internal static void Register()
+    {
+        ItemQueryResolver.Register(ItemQuery_PET_ADOPTION, PET_ADOPTION);
+    }
+
+    internal static IEnumerable<ItemQueryResult> PET_ADOPTION(
+        string key,
+        string arguments,
+        ItemQueryContext context,
+        bool avoidRepeat,
+        HashSet<string> avoidItemIds,
+        Action<string, string> logError
+    )
+    {
+        string[] args = ArgUtility.SplitBySpaceQuoteAware(arguments);
+        if (
+            !ArgUtility.TryGetOptional(
+                args,
+                0,
+                out string petId,
+                out string error,
+                defaultValue: "T",
+                allowBlank: false,
+                name: "string petId"
+            )
+            || !ArgUtility.TryGetOptional(
+                args,
+                1,
+                out string breedId,
+                out error,
+                defaultValue: "T",
+                allowBlank: false,
+                name: "string breedId"
+            )
+            || !ArgUtility.TryGetOptionalBool(
+                args,
+                2,
+                out bool ignoreBasePrice,
+                out error,
+                defaultValue: false,
+                name: "bool ignoreBasePrice"
+            )
+            || !ArgUtility.TryGetOptionalBool(
+                args,
+                3,
+                out bool ignoreCanBeAdoptedFromMarnie,
+                out error,
+                defaultValue: false,
+                name: "bool ignoreCanBeAdoptedFromMarnie"
+            )
+        )
+        {
+            ModEntry.Log(error, LogLevel.Error);
+            return [];
+        }
+        IEnumerable<KeyValuePair<string, PetData>> searchPets;
+        if (petId == "T")
+        {
+            searchPets = Game1.petData;
+        }
+        else
+        {
+            if (Game1.petData.TryGetValue(petId, out PetData? petDataSpecific))
+            {
+                searchPets = [new(petId, petDataSpecific)];
+            }
+            else
+            {
+                ModEntry.Log($"Invalid pet type '{petId}'", LogLevel.Error);
+                return [];
+            }
+        }
+
+        List<ItemQueryResult> list = [];
+        foreach (KeyValuePair<string, PetData> petDatum in searchPets)
+        {
+            foreach (PetBreed breed in petDatum.Value.Breeds)
+            {
+                if (!breed.CanBeAdoptedFromMarnie && !ignoreCanBeAdoptedFromMarnie)
+                {
+                    continue;
+                }
+                if (breedId != "T" && breedId != breed.Id)
+                {
+                    continue;
+                }
+
+                ItemQueryResult result = new(new PetLicense() { Name = petDatum.Key + "|" + breed.Id });
+                if (!ignoreBasePrice)
+                {
+                    result.OverrideBasePrice = breed.AdoptionPrice;
+                }
+                list.Add(result);
+            }
+        }
+        return list;
+    }
+}