+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - WIP
+
+### Added
+
+- Item query mushymato.LivestockBazaar_PET_ADOPTION for filtering for specific pet/breed
+
 ## [1.3.0] - 2025-07-04
 
 ### Added